Fixed installation on Windows
Due fact that this ``xmlrpc-c-config`` is a bash script this does not make much sense on a generic Windows installation. Things are getting even problematic when it comes to the CMake find script which relies on the fact this script is available. To fix this issue in my opinion a CMake config set-up is needed.

Signed-off-by: Christian Ege <christian.ege@ifm.com>